+++ cola/widgets/status.py
@@ -123,8 +123,10 @@ class StatusTreeWidget(QtWidgets.QTreeWidget):
         self.setUniformRowHeights(True)
         self.setAnimated(True)
         self.setRootIsDecorated(False)
-        self.setDragEnabled(True)
         self.setAutoScroll(False)
+        self.setDragEnabled(True)
+        self.setDragDropMode(QtWidgets.QAbstractItemView.DragOnly)
+        self._shift_drag = False
 
         if not prefs.status_indent(context):
             self.setIndentation(0)
@@ -1187,15 +1189,21 @@ class StatusTreeWidget(QtWidgets.QTreeWidget):
         else:
             self.select_by_index(0)
 
+    def mousePressEvent(self, event):
+        """Keep track of whether to drag URLs or just text"""
+        if event.buttons() == Qt.LeftButton:
+            self._shift_drag = event.modifiers() == Qt.ShiftModifier
+        return super(StatusTreeWidget, self).mousePressEvent(event)
+
     def mimeData(self, items):
         """Return a list of absolute-path URLs"""
-        context = self.context
         paths = qtutils.paths_from_items(items, item_filter=_item_filter)
-        return qtutils.mimedata_from_paths(context, paths)
+        include_urls = not self._shift_drag
+        return qtutils.mimedata_from_paths(paths, include_urls=include_urls)
 
     # pylint: disable=no-self-use
     def mimeTypes(self):
-        return qtutils.path_mimetypes()
+        return qtutils.path_mimetypes(include_urls=not self._shift_drag)
 
 
 def _item_filter(item):